# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Achievement:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      file_proof:
        description: File bukti, wajib diisi sebelum status PENDING
        type: string
      id:
        type: string
      level:
        type: string
      points:
        type: integer
      reason:
        description: Alasan penolakan, hanya diisi jika Status = StatusRejected
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.AchievementStatus'
        description: Status Workflow
      student:
        allOf:
        - $ref: '#/definitions/model.Student'
        description: Asumsi model Student dan User sudah terdefinisikan.
      student_id:
        description: Foreign Key ke Student (Pemilik Prestasi)
        type: string
      title:
        type: string
      updated_at:
        type: string
      validator:
        allOf:
        - $ref: '#/definitions/model.Lecturer'
        description: Asumsi model Lecturer sudah terdefinisikan.
      validator_id:
        description: Validator (Dosen Wali/Admin)
        type: string
    type: object
  model.AchievementCreateRequest:
    properties:
      category:
        type: string
      description:
        type: string
      level:
        type: string
      title:
        type: string
    required:
    - category
    - level
    - title
    type: object
...
  model.AchievementStatus:
    enum:
    - DRAFT
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-comments:
      StatusApproved: Status disetujui oleh Dosen Wali/Validator
      StatusDraft: Status awal, bisa diubah/hapus oleh Mahasiswa
      StatusPending: Status setelah di-submit, menunggu verifikasi
      StatusRejected: Status ditolak, bisa di-update/re-submit oleh Mahasiswa
    x-enum-descriptions:
    - Status awal, bisa diubah/hapus oleh Mahasiswa
    - Status setelah di-submit, menunggu verifikasi
    - Status disetujui oleh Dosen Wali/Validator
    - Status ditolak, bisa di-update/re-submit oleh Mahasiswa
    x-enum-varnames:
    - StatusDraft
    - StatusPending
    - StatusApproved
    - StatusRejected
...
  model.Lecturer:
    properties:
      created_at:
        type: string
      department:
        type: string
      id:
        type: string
      lecturer_id:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: string
    type: object
...
  model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      resource:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        description: Gunakan nama package jika Permission berada di package berbeda
        items:
          $ref: '#/definitions/model.Permission'
        type: array
    type: object
...
  model.Student:
    properties:
      academic_year:
        type: string
      advisor:
        $ref: '#/definitions/model.Lecturer'
      advisor_id:
        type: string
      createdAt:
        type: string
      id:
        type: string
      nim:
        type: string
      program_study:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/model.Role'
      role_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@unair.ac.id
    name: API Support
  description: API untuk Sistem Pelaporan Prestasi Mahasiswa
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Student Achievement API
  version: "1.0"
paths:
  /api/v1/achievements:
    get:
      consumes:
      - application/json
      description: 'Mahasiswa: List achievements sendiri (read_own). Dosen/Admin:
        List semua (read_list).'
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Achievement'
                type: array
              status:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
      summary: List Achievements (Filtered by Role)
      tags:
      - Achievements
    post:
      consumes:
      - application/json
      description: Mahasiswa dapat membuat prestasi baru (status DRAFT).
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Prestasi Baru
        in: body
        name: achievement
        required: true
        schema:
          $ref: '#/definitions/model.AchievementCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Achievement'
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
      summary: Create New Achievement
      tags:
      - Achievements
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"